% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gd_loader}
\alias{gd_loader}
\title{Google drive loader}
\usage{
gd_loader(gdDir, subDir, mainDir)
}
\arguments{
\item{gdDir}{}

\item{subDir}{}

\item{mainDir}{}
}
\value{
None
}
\description{
For users with access to the Google Drive holding the data, this function will download the necessary data using the directory structure used in the /analysis/procedures/ files.
}
\examples{
\dontrun{
gd_loader()
}
}
