% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vsurf_bb}
\alias{vsurf_bb}
\title{Title}
\usage{
vsurf_bb(file_x, file_y, ntree = 100, mtry = 2, nfor_thres = 50,
  nmin = 1, nfor_interp = 25, nsd = 1, nfor_pred = 25,
  parallel = "true", ncores = 15, clusterType = "FORK",
  save_template = FALSE, seed = 627, name = "", username = "largesi",
  token = "1ba80d8d-15ea-4ff1-9420-f8c97ea15d40-843339462")
}
\arguments{
\item{token}{}
}
\description{
Title
}
